Turn renv sandbox off by default
From #8. Note that we do not install renv by default. That can be done by the
user But this setting makes sense for all usages of the container I can imagine
and avoids that the user has to set it in .Renviron or .bash_profile. However,
RStudio server will probably not respect this, at least it does not load user
environment variables.